기본적인 소행성의 디스트로이 로직 구현 특정 거리 도달시 파괴 소행성 컴포넌트의 변수 SpawnDistance가 우주 크기의 반지름을 가져오도록 설정하면 됨

+++ Source/TeamSpaceProject/Private/CJH/Stability/ASManager.cpp
@@ -33,7 +33,7 @@ void AASManager::Tick(float DeltaTime)
 
 void AASManager::Artifical_Satellite_Core_Spawn()
 {
-	float Spawn_Num = FMath::RandRange(min_Spawn, max_Spawn);
+	int Spawn_Num = FMath::RandRange(min_Spawn, max_Spawn);
 
 	AActor* Center_Actor = UGameplayStatics::GetActorOfClass(GetWorld(), Center);
 	if (!Center_Actor)
